extract useEditItems
